cleanup: unify 3 different pdf loaders, rename PagedPDFSplitter to PyPDFLoader
`OnlinePDFLoader` and `PagedPDFSplitter` lived separate from the rest of the pdf loaders.

Because they're all similar, moving to `pdy.py` and the same docs/examples page makes sense.

Additionally, PagedPDFSplitter naming doesn't match the pattern the rest of the loaders follow,
so I renamed to `PyPDFLoader` and had it inherit from BasePDFLoader so it can now
load from remote file sources.
