+++ tests/integration_tests/test_pdf_pagesplitter.py
@@ -1,7 +1,7 @@
 """Test splitting with page numbers included."""
 import os
 
-from langchain.document_loaders import PagedPDFSplitter
+from langchain.document_loaders import PyPDFLoader
 from langchain.embeddings.openai import OpenAIEmbeddings
 from langchain.vectorstores import FAISS
 
@@ -9,7 +9,7 @@ from langchain.vectorstores import FAISS
 def test_pdf_pagesplitter() -> None:
     """Test splitting with page numbers included."""
     script_dir = os.path.dirname(__file__)
-    loader = PagedPDFSplitter(os.path.join(script_dir, "examples/hello.pdf"))
+    loader = PyPDFLoader(os.path.join(script_dir, "examples/hello.pdf"))
     docs = loader.load()
     assert "page" in docs[0].metadata
     assert "source" in docs[0].metadata
